# src/main/java/com/bank/islamic/paymentexecution/api/ServiceDomainController.java
package com.bank.islamic.paymentexecution.api;

import com.bank.islamic.paymentexecution.model.ControlRecord;
import com.bank.islamic.paymentexecution.service.ControlRecordStore;
import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.Collection;
import java.util.Map;

/**
 * BIAN semantic API for the "Payment Execution" service domain.
 *
 * Endpoints follow the BIAN action-term style:
 *   GET  /v1/service-domain                          → who am I (SD metadata)
 *   POST /v1/payment-transaction-procedure/initiate                    → Initiate a control record
 *   GET  /v1/payment-transaction-procedure                             → Retrieve (list)
 *   GET  /v1/payment-transaction-procedure/{crId}/retrieve             → Retrieve (single)
 *   PUT  /v1/payment-transaction-procedure/{crId}/update               → Update
 *   PUT  /v1/payment-transaction-procedure/{crId}/control              → Control (suspend|resume|terminate)
 */
@RestController
@RequestMapping("/v1")
public class ServiceDomainController {

    private final ControlRecordStore store;

    public ServiceDomainController(ControlRecordStore store) {
        this.store = store;
    }

    @GetMapping("/service-domain")
    public Map<String, String> serviceDomain() {
        return Map.of(
                "serviceDomain", "Payment Execution",
                "businessArea", "Takaful and Shared Operations",
                "businessDomain", "Islamic Cards and Payments",
                "functionalPattern", "Process",
                "assetType", "Payment Transaction",
                "controlRecord", "Payment Transaction Procedure",
                "version", "0.1.0",
                "phase", "1-shallow"
        );
    }

    @PostMapping("/payment-transaction-procedure/initiate")
    @CircuitBreaker(name = "serviceDomain")
    public ResponseEntity<ControlRecord> initiate(@RequestBody(required = false) Map<String, Object> properties) {
        return ResponseEntity.status(HttpStatus.CREATED).body(store.initiate(properties));
    }

    @GetMapping("/payment-transaction-procedure")
    public Collection<ControlRecord> list() {
        return store.list();
    }

    @GetMapping("/payment-transaction-procedure/{crId}/retrieve")
    public ResponseEntity<ControlRecord> retrieve(@PathVariable String crId) {
        return store.retrieve(crId)
                .map(ResponseEntity::ok)
                .orElse(ResponseEntity.notFound().build());
    }

    @PutMapping("/payment-transaction-procedure/{crId}/update")
    public ResponseEntity<ControlRecord> update(@PathVariable String crId,
                                                @RequestBody Map<String, Object> properties) {
        return store.update(crId, properties)
                .map(ResponseEntity::ok)
                .orElse(ResponseEntity.notFound().build());
    }

    @PutMapping("/payment-transaction-procedure/{crId}/control")
    public ResponseEntity<?> control(@PathVariable String crId,
                                     @RequestBody Map<String, String> body) {
        try {
            return store.control(crId, body.get("action"))
                    .<ResponseEntity<?>>map(ResponseEntity::ok)
                    .orElse(ResponseEntity.notFound().build());
        } catch (IllegalArgumentException e) {
            return ResponseEntity.badRequest().body(Map.of("error", e.getMessage()));
        }
    }
}